refactor: Improve naming of classic identifier conversion in the code
